Mlplan meka/bugfix/not executable (#200)
* Fixed bug that MEKA classifiers could not be executed.

* Resolved compile issue.

* Made GUI events run again for Windows machines.

* Added generic transposition method. However, primitives are not
supported by that.

* Removed command line output.

+++ JAICore/jaicore-ml-meka/src/main/java/ai/libs/jaicore/ml/classification/multilabel/dataset/MekaInstance.java
@@ -28,8 +28,8 @@ public class MekaInstance extends ElementDecorator<Instance> implements IMekaIns
 	}
 
 	@Override
-	public double[] getLabel() {
-		double[] labels = new double[this.getElement().classIndex()];
+	public int[] getLabel() {
+		int[] labels = new int[this.getElement().classIndex()];
 		IntStream.range(0, this.getNumLabels()).forEach(x -> labels[x] = Integer.parseInt(this.getElement().attribute(x).value((int) this.getElement().value(x))));
 		return labels;
 	}
